requirements.txt
aenum==3.1.15
agenta==0.17.5
aiofiles==22.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
altair==4.2.2
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
assemblyai==0.28.1
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
build==1.2.1
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
Chroma==0.2.0
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
croniter==1.3.7
cryptography==41.0.6
dask==2024.6.2
dataclasses-json==0.6.7
datadog==0.44.0
debugpy==1.8.1
decorator==5.1.1
deepdiff==7.0.1
deepgram-sdk==3.3.0
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
elevenlabs==1.3.1
email_validator==2.1.2
emoji==2.12.1
entrypoints==0.4
executing==2.0.1
Faker==4.14.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.20.0
filelock==3.15.1
filetype==1.2.0
Flask==3.0.3
flatbuffers==24.3.25
freezegun==1.2.2
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.30.0
googleapis-common-protos==1.63.1
great-expectations==0.18.12
greenlet==3.0.3
groq==0.9.0
grpcio==1.64.1
h11==0.14.0
helper_utils==0.0.8
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflection==0.5.1
intel-openmp==2021.4.0
ipdb==0.13.13
ipykernel==6.15.0
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-server==1.23.5
jupyter-server-proxy==3.2.3
jupyter_client==7.4.4
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.11
kubernetes==30.1.0
langchain==0.2.6
langchain-chroma==0.1.1
langchain-community==0.2.6
langchain-core==0.2.11
langchain-groq==0.1.5
langchain-huggingface==0.0.3
langchain-nomic==0.1.2
langchain-openai==0.1.8
langchain-pinecone==0.1.1
langchain-text-splitters==0.2.1
langdetect==1.0.9
langsmith==0.1.77
ldap3==2.9.1
Levenshtein==0.25.1
llama-index-core==0.10.46
llama-parse==0.4.4
llamaindex-py-client==0.1.19
locket==1.0.0
loguru==0.7.2
lxml==5.2.2
mage-ai==0.9.72
makefun==1.15.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.0.2
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclassic==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
newrelic==8.8.0
nltk==3.8.1
nomic==3.0.31
notebook==6.5.7
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.0
openai==1.34.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
ordered-set==4.1.0
orjson==3.10.4
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pdfminer.six==20231228
pdfplumber==0.11.1
pep517==0.13.1
pickleshare==0.7.5
pillow==10.3.0
pinecone==4.0.0
pinecone-client==3.2.2
pinecone-notebooks==0.1.1
platformdirs==4.2.2
playsound==1.3.0
polars==0.19.1
posthog==3.5.0
prometheus_client==0.20.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.8
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
PyGithub==1.59.0
Pygments==2.18.0
PyJWT==2.6.0
pymongo==4.8.0
PyNaCl==1.5.0
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-Levenshtein==0.25.1
python-magic==0.4.27
python-multipart==0.0.9
python-weather==2.0.3
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
questionary==2.0.1
rapidfuzz==3.9.3
redis==5.0.6
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruamel.yaml==0.17.17
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
selenium==4.22.0
Send2Trash==1.8.3
sentence-transformers==3.0.1
sentry-sdk==1.19.1
shellingham==1.5.4
simpervisor==1.0.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.7
soupsieve==2.5
SQLAlchemy==1.4.52
sqlglot==25.3.0
sqlglotrs==0.2.7
stack-data==0.6.3
starlette==0.37.2
sympy==1.12.1
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.3.0
terminado==0.17.1
text-unidecode==1.3
thefuzz==0.19.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
tornado==6.3.3
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
trio==0.26.0
trio-websocket==0.11.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
unstructured==0.14.6
unstructured-client==0.23.7
urllib3==2.2.1
uvicorn==0.30.1
watchdog==4.0.0
watchfiles==0.22.0
wcwidth==0.2.13
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
widgetsnbextension==4.0.11
win32-setctime==1.1.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
